# contracts/openapi/gateway.yaml
openapi: 3.1.0
info:
  title: Swarm Blackjack - API Gateway
  description: |
    Single external entry point for all client traffic.
    TLS termination, JWT validation, request routing, rate limiting.
    Nothing else is externally exposed.
  version: 0.1.0
  contact:
    name: Swarm Blackjack PoC

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: game
    description: Game operations (proxied to game-state service)
  - name: chat
    description: Chat operations (proxied to chat-service via WebSocket)
  - name: auth
    description: Auth operations (proxied to auth-service)
  - name: bank
    description: Financial operations (proxied to bank-service)
  - name: observability
    description: Live swarm activity feed

paths:
  /health:
    get:
      summary: Gateway health check
      tags: [observability]
      responses:
        '200':
          description: Gateway is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /api/game/{tableId}/stream:
    get:
      summary: Subscribe to game state via Server-Sent Events
      description: |
        SSE stream delivering real-time game state updates.
        Client receives events as game progresses through phases.
        Connection is long-lived — reconnect on disconnect.
      tags: [game]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TableId'
      responses:
        '200':
          description: SSE stream established
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/GameStateEvent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/game/{tableId}/action:
    post:
      summary: Submit a player action
      description: |
        Player actions are discrete HTTP POST requests, not WebSocket messages.
        Each action is independently authenticated and policy-checked.
      tags: [game]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TableId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlayerAction'
      responses:
        '202':
          description: Action accepted, state update will arrive via SSE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionAccepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Action not valid in current game phase
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/game/tables:
    get:
      summary: List available tables
      tags: [game]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Available tables
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TableSummary'

  /api/auth/register:
    post:
      summary: Begin passkey registration
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationRequest'
      responses:
        '200':
          description: Registration options (WebAuthn ceremony begins)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationOptions'

  /api/auth/login:
    post:
      summary: Begin passkey authentication
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication options (WebAuthn ceremony begins)
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/auth/refresh:
    post:
      summary: Refresh JWT using Redis session token
      tags: [auth]
      responses:
        '200':
          description: New JWT issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/bank/balance:
    get:
      summary: Get player chip balance
      tags: [bank]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'

  /events:
    get:
      summary: Observability SSE feed — live swarm activity
      description: |
        SSE stream of all inter-service calls, routed through the gateway.
        Used by the Observability Dashboard to visualize swarm activity.
        Shows service name, endpoint, latency, and response status for every call.
      tags: [observability]
      responses:
        '200':
          description: Observability event stream
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/ObservabilityEvent'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived JWT (15 min). Refresh via /api/auth/refresh.

  parameters:
    TableId:
      name: tableId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique table identifier

  schemas:
    Card:
      type: object
      required: [suit, rank]
      properties:
        suit:
          type: string
          enum: [hearts, diamonds, clubs, spades, hidden]
        rank:
          type: string
          enum: [A, '2', '3', '4', '5', '6', '7', '8', '9', '10', J, Q, K, hidden]

    PlayerState:
      type: object
      required: [id, name, chips, hand, handValue, status]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        chips:
          type: integer
          description: Current chip count
        currentBet:
          type: integer
          description: Bet placed this round
        hand:
          type: array
          items:
            $ref: '#/components/schemas/Card'
        handValue:
          type: integer
          minimum: 0
          maximum: 30
        isSoftHand:
          type: boolean
        status:
          type: string
          enum: [waiting, betting, playing, standing, bust, blackjack, won, lost, push]

    DealerState:
      type: object
      required: [hand, handValue]
      properties:
        hand:
          type: array
          items:
            $ref: '#/components/schemas/Card'
        handValue:
          type: integer
        isRevealed:
          type: boolean
          description: False during player turns (hole card hidden)

    GameStateEvent:
      type: object
      description: SSE event payload — game state snapshot
      required: [type, tableId, phase, players, dealer, timestamp]
      properties:
        type:
          type: string
          enum: [game_state, phase_change, player_joined, player_left, error]
        tableId:
          type: string
          format: uuid
        phase:
          type: string
          enum: [waiting, betting, dealing, player_turn, dealer_turn, payout, complete]
        players:
          type: array
          items:
            $ref: '#/components/schemas/PlayerState'
        dealer:
          $ref: '#/components/schemas/DealerState'
        activePlayerId:
          type: string
          format: uuid
          nullable: true
        handledBy:
          type: string
          description: Service container that generated this state update
        timestamp:
          type: string
          format: date-time

    PlayerAction:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [bet, hit, stand, double, split, insurance]
        amount:
          type: integer
          description: Required for bet action
          minimum: 1

    ActionAccepted:
      type: object
      properties:
        accepted: 
          type: boolean
        message:
          type: string

    TableSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        playerCount:
          type: integer
        maxPlayers:
          type: integer
        phase:
          type: string
        minBet:
          type: integer
        maxBet:
          type: integer

    RegistrationRequest:
      type: object
      required: [email, displayName]
      properties:
        email:
          type: string
          format: email
        displayName:
          type: string

    RegistrationOptions:
      type: object
      description: WebAuthn PublicKeyCredentialCreationOptions

    LoginRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email

    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        expiresIn:
          type: integer
          description: Seconds until expiry

    BalanceResponse:
      type: object
      properties:
        playerId:
          type: string
          format: uuid
        chips:
          type: integer
        currency:
          type: string
          default: chips

    ObservabilityEvent:
      type: object
      description: SSE event for observability dashboard
      properties:
        id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        caller:
          type: string
          description: Service making the call
        callee:
          type: string
          description: Service being called
        method:
          type: string
          enum: [GET, POST, PUT, DELETE, SSE, WS]
        path:
          type: string
        statusCode:
          type: integer
        latencyMs:
          type: integer
        protocol:
          type: string
          enum: [http, https, sse, websocket, mtls]

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, degraded, unhealthy]
        service:
          type: string
        version:
          type: string
        upstreamServices:
          type: object
          additionalProperties:
            type: string

    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string

  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: Invalid request body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'